ecommerce_api/requirements.txt
django>=5.0
djangorestframework
djangorestframework-simplejwt
django-filter
pillow
psycopg[binary]
dj-database-url
python-dotenv
whitenoise
gunicorn
